add lpNorm implement

+++ mindspore/core/utils/convert_utils_base.h
@@ -101,7 +101,7 @@ inline float SizeToFloat(size_t v) { return static_cast<float>(v); }
 
 inline double LongToDouble(int64_t v) { return static_cast<double>(v); }
 
-inline double LongToFloat(int64_t v) { return static_cast<float>(v); }
+inline float LongToFloat(int64_t v) { return static_cast<float>(v); }
 
 inline double FloatToDouble(float v) { return static_cast<double>(v); }
